update phân trang theo mã thương hiệu, danh mục và sắp xếp kéo thả thứ tự xuất hiện của bannner ,quản lý tồn kho và chi nhánh cửa hàng và kho tổng

+++ DATN/src/main/java/com/example/DATN/services/FileStorageService.java
@@ -109,7 +109,10 @@ public class FileStorageService {
 
     public void deleteFile(String filename) {
         try {
-            Path filePath = this.storageFolder.resolve(filename).normalize().toAbsolutePath();
+            Path filePath = findFileInAllFolders(filename);
+            if(filePath==null){
+                throw new ApplicationException(ErrorCode.FILE_EMPTY);
+            }
             Files.deleteIfExists(filePath);
         } catch (IOException e) {
             throw new RuntimeException("Failed to delete file: " + filename, e);
